Ensure that `Capybara.server_host` setting isn't inherited
If the user of fake_braintree also uses Capybara for their acceptance
tests, FakeBraintree::Server will have erroneously inherited the
`server_host` setting that is not meant for it.

Ensure that we bind to 127.0.0.1 always. This needs to bump Capybara
version requirement to v2.2.0, which is when Capybara::Server started
receiving the host argument.