Bugfix: Ignore empty discount/addon arrays
The Braintree API is just fine with accepting an empty array – so FakeBraintree
should be fine with this as well.

Example:

Braintree::Subscription.update(subscription_id, discounts: {
	add: [],
	update: [{ existing_id: some_id, amount: some_amount }]
})

This previously failed because FakeBraintree would stop at the `add` key since
it can be found and isn't nil. After this patch, FakeBraintree continues on the
empty array and correctly updates the discount.